+++ catalog/category.models.ts
@@ -1,0 +1,27 @@
+import {CreationOptional, DataTypes, InferAttributes, InferCreationAttributes, Model} from 'sequelize'
+import {sequelize} from '../utils/postgresqlConnect'
+import {Product} from './product.model'
+
+interface CategoryModel extends Model<InferAttributes<CategoryModel>, InferCreationAttributes<CategoryModel>> {
+    id: CreationOptional<number>
+    name: string;
+    description: string
+
+}
+
+export const Category = sequelize.define<CategoryModel>('Category', {
+    id: {
+        type: DataTypes.INTEGER,
+        autoIncrement: true,
+        primaryKey: true
+    },
+    name: {
+        type: DataTypes.STRING,
+    },
+    description: {
+        type: DataTypes.STRING,
+    }
+})
+
+Category.hasMany(Product)
+Product.belongsTo(Category)
